Adds a test that actually asserts that there is no more write after close
